+++ src/main/java/ru/shapovalov/service/AccountService.java
@@ -1,10 +1,10 @@
 package ru.shapovalov.service;
 
-import ru.shapovalov.converter.AccountToAccountDtoConverter;
-import ru.shapovalov.entity.Account;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.shapovalov.api.converter.AccountToAccountDtoConverter;
 import ru.shapovalov.dao.AccountDao;
+import ru.shapovalov.entity.Account;
 
 import java.util.List;
 
@@ -16,22 +16,22 @@ public class AccountService {
     private final AccountDao accountDao;
     private final AccountToAccountDtoConverter accountDtoConverter;
 
-    public List<AccountDto> getAll(int userId) {
+    public List<AccountDto> getAll(Long userId) {
         return accountDao.getAllByUserId(userId).stream()
                 .map(accountDtoConverter::convert)
                 .collect(toList());
     }
 
-    public AccountDto create(String accountName, int userId) {
+    public AccountDto create(String accountName, Long userId) {
         Account account = accountDao.insert(accountName, userId);
         return accountDtoConverter.convert(account);
     }
 
-    public boolean edit(int accountId, String newName, int userId) {
+    public boolean edit(Long accountId, String newName, Long userId) {
         return accountDao.edit(accountId, newName, userId);
     }
 
-    public boolean delete(int accountId, int userId) {
+    public boolean delete(Long accountId, Long userId) {
         return accountDao.delete(accountId, userId);
     }
 }
