+++ src/main/java/ru/shapovalov/service/CategoryService.java
@@ -1,10 +1,10 @@
 package ru.shapovalov.service;
 
-import ru.shapovalov.entity.Category;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.shapovalov.converter.CategoryModelToCategoryDtoConverter;
+import ru.shapovalov.api.converter.CategoryModelToCategoryDtoConverter;
 import ru.shapovalov.dao.CategoryDao;
+import ru.shapovalov.entity.Category;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
@@ -20,21 +20,21 @@ public class CategoryService {
     private final CategoryDao categoryDao;
     private final CategoryModelToCategoryDtoConverter categoryDtoConverter;
 
-    public CategoryDto create(String categoryName, int userId) {
+    public CategoryDto create(String categoryName, Long userId) {
         Category category = categoryDao.insert(categoryName, userId);
         return categoryDtoConverter.convert(category);
     }
 
-    public boolean delete(int id, int userId) {
+    public boolean delete(Long id, Long userId) {
         return categoryDao.delete(id, userId);
     }
 
-    public CategoryDto edit(int id, String newCategoryName, int userId) {
+    public CategoryDto edit(Long id, String newCategoryName, Long userId) {
         Category category = categoryDao.edit(id, newCategoryName, userId);
         return categoryDtoConverter.convert(category);
     }
 
-    public Map<String, Long> getResultIncomeInPeriodByCategory(int userId, int days) {
+    public Map<String, Long> getResultIncomeInPeriodByCategory(Long userId, int days) {
         LocalDateTime now = LocalDateTime.now();
         LocalDateTime yesterday = now.minusDays(days);
 
@@ -44,7 +44,7 @@ public class CategoryService {
         return categoryDao.getResultIncomeInPeriodByCategory(userId, startDate, endDate);
     }
 
-    public Map<String, Long> getResultExpenseInPeriodByCategory(int userId, int days) {
+    public Map<String, Long> getResultExpenseInPeriodByCategory(Long userId, int days) {
         LocalDateTime now = LocalDateTime.now();
         LocalDateTime yesterday = now.minusDays(days);
 
@@ -54,7 +54,7 @@ public class CategoryService {
         return categoryDao.getResultExpenseInPeriodByCategory(userId, startDate, endDate);
     }
 
-    public List<CategoryDto> getAll(int userId) {
+    public List<CategoryDto> getAll(Long userId) {
         return categoryDao.getAllByUserId(userId).stream()
                 .map(categoryDtoConverter::convert)
                 .collect(toList());
