+++ src/main/java/ru/shapovalov/service/TransactionDto.java
@@ -8,9 +8,9 @@ import java.sql.Timestamp;
 @Data
 @Accessors(chain = true)
 public class TransactionDto {
-    private int id;
-    private Integer sender;
-    private Integer recipient;
+    private Long id;
+    private Long sender;
+    private Long recipient;
     private int sum;
     private Timestamp createdDate;
 }
